requirements.txt
# Core database drivers
psycopg2-binary>=2.9.0
clickhouse-driver>=0.2.7
influxdb-client>=1.38.0
asyncpg>=0.29.0

# Data processing (Python 3.13 compatible versions)
pandas>=2.2.0
numpy>=1.26.0

# Utilities
colorama>=0.4.6
tabulate>=0.9.0
pyyaml>=6.0.1
psutil>=5.9.6

